% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descriptions}
\alias{descriptions}
\title{View the data sets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 89 rows and 10 columns.
}
\usage{
descriptions
}
\description{
@format Data set of multilayer networks
}
\details{
@source {create}

 @examples
 data(descriptions)
}
\keyword{datasets}
