% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_to_json.R
\name{.df_to_list_of_rows}
\alias{.df_to_list_of_rows}
\title{Convert a data.frame to a list of named lists (one per row)}
\usage{
.df_to_list_of_rows(df)
}
\description{
Convert a data.frame to a list of named lists (one per row)
}
\keyword{internal}
