% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln.R
\docType{package}
\name{emln-package}
\alias{emln}
\alias{emln-package}
\title{emln: Organize, Handle, and Explore Ecological Multilayer Networks}
\description{
Data and analysis of ecological multilayer networks, including standardization of data structures and functions to convert between them. Includes an interactive multilayer network visualizer (beta, paper forthcoming), and a collection of 78 empirical ecological multilayer network datasets. This work was supported by research grant ISF (Israel Science Foundation) 1281/20 to Shai Pilosof. Noa Frydman (2023) \doi{10.1111/2041-210X.14225}. Package code is licensed under GPL-3. Mira, the network visualizer, is licensed under CC-BY-NC-4.0; see inst/MiRA/LICENSE.md for full terms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emln.ecomplab.com/}
  \item Report bugs at \url{https://github.com/Ecological-Complexity-Lab/emln/issues}
}

}
\author{
\strong{Maintainer}: Geut Galai \email{geutg@post.bgu.ac.il}

Authors:
\itemize{
  \item Noa Frydman
  \item Shai Pilosof \email{pilos@bgu.ac.il} [funder]
  \item Shirly Freilikhman
}

}
\keyword{internal}
