% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln10_environment_temporal_parker_huryn_2006.R
\docType{data}
\name{emln10_environment_temporal_parker_huryn_2006}
\alias{emln10_environment_temporal_parker_huryn_2006}
\title{emln10_environment_temporal_parker_huryn_2006}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   10 \tab Environment, Temporal \tab 141 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
