% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln11_environment_thompson_townsend_2004.R
\docType{data}
\name{emln11_environment_thompson_townsend_2004}
\alias{emln11_environment_thompson_townsend_2004}
\title{emln11_environment_thompson_townsend_2004}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   11 \tab Environment \tab 154 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
