% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln12_environment_fryer_1959.R
\docType{data}
\name{emln12_environment_fryer_1959}
\alias{emln12_environment_fryer_1959}
\title{emln12_environment_fryer_1959}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   12 \tab Environment \tab 94 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
