% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln13_multiplex_kefi_2016.R
\docType{data}
\name{emln13_multiplex_kefi_2016}
\alias{emln13_multiplex_kefi_2016}
\title{emln13_multiplex_kefi_2016}
\source{
Dryad

https://datadryad.org/stash/dataset/doi:10.5061/dryad.b4vg0
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   13 \tab Multiplex \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
