% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln14_multiplex_mello_2019.R
\docType{data}
\name{emln14_multiplex_mello_2019}
\alias{emln14_multiplex_mello_2019}
\title{emln14_multiplex_mello_2019}
\source{
Web_of_Life

http://www.web-of-life.es
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   14 \tab Multiplex \tab Seed-Dispersal \tab FALSE \cr
}
}
\keyword{internal}
