% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln15_environment_petipa_1970.R
\docType{data}
\name{emln15_environment_petipa_1970}
\alias{emln15_environment_petipa_1970}
\title{emln15_environment_petipa_1970}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   15 \tab Environment \tab 50 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
