% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln16_multiplex_pocock_2012.R
\docType{data}
\name{emln16_multiplex_pocock_2012}
\alias{emln16_multiplex_pocock_2012}
\title{emln16_multiplex_pocock_2012}
\source{
Dryad

https://datadryad.org/stash/dataset/doi:10.5061/dryad.3s36r118
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   16 \tab Multiplex \tab Multiple \tab FALSE \cr
}
}
\keyword{internal}
