% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln17_multiplex_traveset_2020.R
\docType{data}
\name{emln17_multiplex_traveset_2020}
\alias{emln17_multiplex_traveset_2020}
\title{emln17_multiplex_traveset_2020}
\source{
Figshare

https://doi.org/10.6084/m9.figshare.c.5208546.v1
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   17 \tab Multiplex \tab Multiple \tab FALSE \cr
}
}
\keyword{internal}
