% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln18_perturbation_stewart_sprules_2011.R
\docType{data}
\name{emln18_perturbation_stewart_sprules_2011}
\alias{emln18_perturbation_stewart_sprules_2011}
\title{emln18_perturbation_stewart_sprules_2011}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   18 \tab Perturbation \tab 160 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
