% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln19_spatial_angelini.R
\docType{data}
\name{emln19_spatial_angelini}
\alias{emln19_spatial_angelini}
\title{emln19_spatial_angelini}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   19 \tab Spatial \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
