% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln1_environment_schleuning_2010.R
\docType{data}
\name{emln1_environment_schleuning_2010}
\alias{emln1_environment_schleuning_2010}
\title{emln1_environment_schleuning_2010}
\source{
Interaction_Web_DataBase

http://www.ecologia.ib.usp.br/iwdb/html/schleuning-et-al-2010.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   1 \tab Environment \tab Seed-Dispersal \tab FALSE \cr
}
}
\keyword{internal}
