% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln20_spatial_angelini_2013.R
\docType{data}
\name{emln20_spatial_angelini_2013}
\alias{emln20_spatial_angelini_2013}
\title{emln20_spatial_angelini_2013}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   20 \tab Spatial \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
