% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln21_spatial_arroyo_1982.R
\docType{data}
\name{emln21_spatial_arroyo_1982}
\alias{emln21_spatial_arroyo_1982}
\title{emln21_spatial_arroyo_1982}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php

http://www.ecologia.ib.usp.br/iwdb/html/arroyo_1982.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   21 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
