% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln22_spatial_beaver_1985.R
\docType{data}
\name{emln22_spatial_beaver_1985}
\alias{emln22_spatial_beaver_1985}
\title{emln22_spatial_beaver_1985}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   22 \tab Spatial \tab 86 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
