% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln23_spatial_canadian_fish_parasite.R
\docType{data}
\name{emln23_spatial_canadian_fish_parasite}
\alias{emln23_spatial_canadian_fish_parasite}
\title{emln23_spatial_canadian_fish_parasite}
\source{
Interaction_Web_DataBase

http://www.ecologia.ib.usp.br/iwdb/html/canadian_fish.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   23 \tab Spatial \tab Host-Parasite \tab TRUE \cr
}
}
\keyword{internal}
