% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln24_spatial_carlo_2003.R
\docType{data}
\name{emln24_spatial_carlo_2003}
\alias{emln24_spatial_carlo_2003}
\title{emln24_spatial_carlo_2003}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   24 \tab Spatial \tab Seed-Dispersal \tab FALSE \cr
}
}
\keyword{internal}
