% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln25_spatial_carstensen_2014.R
\docType{data}
\name{emln25_spatial_carstensen_2014}
\alias{emln25_spatial_carstensen_2014}
\title{emln25_spatial_carstensen_2014}
\source{
figshare

https://figshare.com/articles/dataset/\emph{Beta_Diversity_of_Plant_Pollinator_Networks_and_the_Spatial_Turnover_of_Pairwise_Interactions}/1234708
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   25 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
