% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln27_spatial_dexter_1947.R
\docType{data}
\name{emln27_spatial_dexter_1947}
\alias{emln27_spatial_dexter_1947}
\title{emln27_spatial_dexter_1947}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   27 \tab Spatial \tab 54 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
