% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln28_spatial_dicks_2002.R
\docType{data}
\name{emln28_spatial_dicks_2002}
\alias{emln28_spatial_dicks_2002}
\title{emln28_spatial_dicks_2002}
\source{
Web_of_Life

http://www.web-of-life.es/map.php?type=6
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   28 \tab Spatial \tab Pollination \tab TRUE \cr
}
}
\keyword{internal}
