% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln29_spatial_dupont_olesen_2009.R
\docType{data}
\name{emln29_spatial_dupont_olesen_2009}
\alias{emln29_spatial_dupont_olesen_2009}
\title{emln29_spatial_dupont_olesen_2009}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   29 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
