% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln2_environment_benadi_jae_2013.R
\docType{data}
\name{emln2_environment_benadi_jae_2013}
\alias{emln2_environment_benadi_jae_2013}
\title{emln2_environment_benadi_jae_2013}
\source{
Dryad

https://datadryad.org/stash/dataset/doi:10.5061/dryad.8mn44
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   2 \tab Environment \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
