% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln30_spatial_elberling_1999.R
\docType{data}
\name{emln30_spatial_elberling_1999}
\alias{emln30_spatial_elberling_1999}
\title{emln30_spatial_elberling_1999}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php?type=6

http://www.ecologia.ib.usp.br/iwdb/
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   30 \tab Spatial \tab Pollination \tab TRUE \cr
}
}
\keyword{internal}
