% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln31_spatial_environment_cornaby_1974.R
\docType{data}
\name{emln31_spatial_environment_cornaby_1974}
\alias{emln31_spatial_environment_cornaby_1974}
\title{emln31_spatial_environment_cornaby_1974}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   31 \tab Spatial    , Environment \tab 90 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
