% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln32_spatial_environment_mckinnerney_1978.R
\docType{data}
\name{emln32_spatial_environment_mckinnerney_1978}
\alias{emln32_spatial_environment_mckinnerney_1978}
\title{emln32_spatial_environment_mckinnerney_1978}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   32 \tab Spatial    , Environment \tab 148 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
