% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln33_spatial_hadfield_2013.R
\docType{data}
\name{emln33_spatial_hadfield_2013}
\alias{emln33_spatial_hadfield_2013}
\title{emln33_spatial_hadfield_2013}
\source{
Web_of_Life

Dryad

http://www.web-of-life.es/map.php

https://datadryad.org/stash/dataset/doi:10.5061/dryad.jf3tj
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   33 \tab Spatial \tab Host-Parasite \tab TRUE \cr
}
}
\keyword{internal}
