% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln34_spatial_havens_1992.R
\docType{data}
\name{emln34_spatial_havens_1992}
\alias{emln34_spatial_havens_1992}
\title{emln34_spatial_havens_1992}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   34 \tab Spatial \tab 15 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
