% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln35_spatial_ingversen_2006.R
\docType{data}
\name{emln35_spatial_ingversen_2006}
\alias{emln35_spatial_ingversen_2006}
\title{emln35_spatial_ingversen_2006}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   35 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
