% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln36_spatial_joern_1979.R
\docType{data}
\name{emln36_spatial_joern_1979}
\alias{emln36_spatial_joern_1979}
\title{emln36_spatial_joern_1979}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php?type=6

http://www.ecologia.ib.usp.br/iwdb/html/joern_1979.htm
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   36 \tab Spatial \tab Plant-Herbivore \tab TRUE \cr
}
}
\keyword{internal}
