% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln37_spatial_jordano_unpubl.R
\docType{data}
\name{emln37_spatial_jordano_unpubl}
\alias{emln37_spatial_jordano_unpubl}
\title{emln37_spatial_jordano_unpubl}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   37 \tab Spatial \tab Seed-Dispersal \tab FALSE \cr
}
}
\keyword{internal}
