% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln38_spatial_kyoto.R
\docType{data}
\name{emln38_spatial_kyoto}
\alias{emln38_spatial_kyoto}
\title{emln38_spatial_kyoto}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php

http://www.ecologia.ib.usp.br/iwdb/resources.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   38 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
