% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln39_spatial_leather_1991.R
\docType{data}
\name{emln39_spatial_leather_1991}
\alias{emln39_spatial_leather_1991}
\title{emln39_spatial_leather_1991}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php

http://www.ecologia.ib.usp.br/iwdb/html/leather_1991.htm
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   39 \tab Spatial \tab Plant-Herbivore \tab TRUE \cr
}
}
\keyword{internal}
