% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln3_environment_bird_1930.R
\docType{data}
\name{emln3_environment_bird_1930}
\alias{emln3_environment_bird_1930}
\title{emln3_environment_bird_1930}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   3 \tab Environment \tab 87 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
