% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln40_spatial_medan_2002.R
\docType{data}
\name{emln40_spatial_medan_2002}
\alias{emln40_spatial_medan_2002}
\title{emln40_spatial_medan_2002}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.ecologia.ib.usp.br/iwdb/html/medan_et_al_2002.html

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   40 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
