% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln41_spatial_melian_2015.R
\docType{data}
\name{emln41_spatial_melian_2015}
\alias{emln41_spatial_melian_2015}
\title{emln41_spatial_melian_2015}
\source{
Dryad

https://datadryad.org/stash/dataset/doi:10.5061/dryad.06bb5
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   41 \tab Spatial \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
