% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln42_spatial_menge_sutherland_1976.R
\docType{data}
\name{emln42_spatial_menge_sutherland_1976}
\alias{emln42_spatial_menge_sutherland_1976}
\title{emln42_spatial_menge_sutherland_1976}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   42 \tab Spatial \tab 104 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
