% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln43_spatial_montero_2005.R
\docType{data}
\name{emln43_spatial_montero_2005}
\alias{emln43_spatial_montero_2005}
\title{emln43_spatial_montero_2005}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   43 \tab Spatial \tab Pollination \tab TRUE \cr
}
}
\keyword{internal}
