% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln44_spatial_olesen_2003.R
\docType{data}
\name{emln44_spatial_olesen_2003}
\alias{emln44_spatial_olesen_2003}
\title{emln44_spatial_olesen_2003}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   44 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
