% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln45_spatial_olesen_2002.R
\docType{data}
\name{emln45_spatial_olesen_2002}
\alias{emln45_spatial_olesen_2002}
\title{emln45_spatial_olesen_2002}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   45 \tab Spatial \tab 61 \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
