% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln47_spatial_paine_1963.R
\docType{data}
\name{emln47_spatial_paine_1963}
\alias{emln47_spatial_paine_1963}
\title{emln47_spatial_paine_1963}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   47 \tab Spatial \tab 110 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
