% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/emln48_environment_perturbation_bartomeus_2008.R
\docType{data}
\name{emln48_environment_perturbation_bartomeus_2008}
\alias{emln48_environment_perturbation_bartomeus_2008}
\title{emln48_environment_perturbation_bartomeus_2008}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.ecologia.ib.usp.br/iwdb/html/bartomeus_2008.html

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   48 \tab Environment , Perturbation \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
