% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln49_spatial_perturbation_vazquez_2002.R
\docType{data}
\name{emln49_spatial_perturbation_vazquez_2002}
\alias{emln49_spatial_perturbation_vazquez_2002}
\title{emln49_spatial_perturbation_vazquez_2002}
\source{
Interaction_Web_DataBase

http://www.ecologia.ib.usp.br/iwdb/html/vazquez_2002.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   49 \tab Spatial     , Perturbation \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
