% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln4_environment_hawkins_goeden_1984.R
\docType{data}
\name{emln4_environment_hawkins_goeden_1984}
\alias{emln4_environment_hawkins_goeden_1984}
\title{emln4_environment_hawkins_goeden_1984}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   4 \tab Environment \tab 145 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
