% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln50_spatial_poulin_1999.R
\docType{data}
\name{emln50_spatial_poulin_1999}
\alias{emln50_spatial_poulin_1999}
\title{emln50_spatial_poulin_1999}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php

http://www.ecologia.ib.usp.br/iwdb/resources.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   50 \tab Spatial \tab Seed-Dispersal \tab FALSE \cr
}
}
\keyword{internal}
