% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln51_spatial_primack_1983.R
\docType{data}
\name{emln51_spatial_primack_1983}
\alias{emln51_spatial_primack_1983}
\title{emln51_spatial_primack_1983}
\source{
Web_of_Life

http://www.web-of-life.es
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   51 \tab Spatial \tab Pollination \tab TRUE \cr
}
}
\keyword{internal}
