% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln52_spatial_ricciardi_2010.R
\docType{data}
\name{emln52_spatial_ricciardi_2010}
\alias{emln52_spatial_ricciardi_2010}
\title{emln52_spatial_ricciardi_2010}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   52 \tab Spatial \tab 74 \tab Anemone-Fish \tab FALSE \cr
}
}
\keyword{internal}
