% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln53_spatial_slovakia.R
\docType{data}
\name{emln53_spatial_slovakia}
\alias{emln53_spatial_slovakia}
\title{emln53_spatial_slovakia}
\source{
NA

NA
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   53 \tab Spatial \tab Host-Parasite \tab FALSE \cr
}
}
\keyword{internal}
