% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln54_Spatial_Snow_B.K.Snow_1988.R
\docType{data}
\name{emln54_spatial_snow_b_k_snow_1988}
\alias{emln54_spatial_snow_b_k_snow_1988}
\title{emln54_spatial_snow b_k_&snow_1988}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.ecologia.ib.usp.br/iwdb/html/snow_snow_1971.htm

http://www.web-of-life.es/map.php?type=6
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   54 \tab Spatial \tab Pollination \tab TRUE \cr
}
}
\keyword{internal}
