% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln55_spatial_temporal_alcorlo_2001.R
\docType{data}
\name{emln55_spatial_temporal_alcorlo_2001}
\alias{emln55_spatial_temporal_alcorlo_2001}
\title{emln55_spatial_temporal_alcorlo_2001}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   55 \tab Spatial , Temporal \tab 81 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
