% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln56_spatial_temporal_closs_1994.R
\docType{data}
\name{emln56_spatial_temporal_closs_1994}
\alias{emln56_spatial_temporal_closs_1994}
\title{emln56_spatial_temporal_closs_1994}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   56 \tab Spatial , Temporal \tab 49 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
