% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln57_spatial_temporal_kolpelke_2017.R
\docType{data}
\name{emln57_spatial_temporal_kolpelke_2017}
\alias{emln57_spatial_temporal_kolpelke_2017}
\title{emln57_spatial_temporal_kolpelke_2017}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   57 \tab Spatial , Temporal \tab 6 \tab Host-Parasite \tab FALSE \cr
}
}
\keyword{internal}
