% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln58_spatial_temporal_martins_2020.R
\docType{data}
\name{emln58_spatial_temporal_martins_2020}
\alias{emln58_spatial_temporal_martins_2020}
\title{emln58_spatial_temporal_martins_2020}
\source{
Dryad

https://doi.org/10.5061/dryad.vmcvdncq0
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   58 \tab Spatial , Temporal \tab Plant-Herbivore \tab TRUE \cr
}
}
\keyword{internal}
