% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln59_spatial_thompson_2003.R
\docType{data}
\name{emln59_spatial_thompson_2003}
\alias{emln59_spatial_thompson_2003}
\title{emln59_spatial_thompson_2003}
\source{
Web_of_Life

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   59 \tab Spatial \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
