% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln5_environment_jones_1949.R
\docType{data}
\name{emln5_environment_jones_1949}
\alias{emln5_environment_jones_1949}
\title{emln5_environment_jones_1949}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   5 \tab Environment \tab 99 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
