% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln60_spatial_trojelsgaard_2015.R
\docType{data}
\name{emln60_spatial_trojelsgaard_2015}
\alias{emln60_spatial_trojelsgaard_2015}
\title{emln60_spatial_trojelsgaard_2015}
\source{
Dryad

https://datadryad.org/stash/dataset/doi:10.5061/dryad.76173
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   60 \tab Spatial \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
