% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln62_temporal_angelini_2006.R
\docType{data}
\name{emln62_temporal_angelini_2006}
\alias{emln62_temporal_angelini_2006}
\title{emln62_temporal_angelini_2006}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   62 \tab Temporal \tab 84 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
