% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln64_temporal_caradona_2020.R
\docType{data}
\name{emln64_temporal_caradona_2020}
\alias{emln64_temporal_caradona_2020}
\title{emln64_temporal_caradona_2020}
\source{
EDI_data_portal

https://portal.edirepository.org/nis/mapbrowse?packageid=edi.512.1
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   64 \tab Temporal \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
