% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln65_temporal_chacoff_2020.R
\docType{data}
\name{emln65_temporal_chacoff_2020}
\alias{emln65_temporal_chacoff_2020}
\title{emln65_temporal_chacoff_2020}
\source{
Dryad

https://datadryad.org/stash/dataset/doi:10.5061/dryad.8cz8w9gm1
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   65 \tab Temporal \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
