% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln66_temporal_cohen_2003.R
\docType{data}
\name{emln66_temporal_cohen_2003}
\alias{emln66_temporal_cohen_2003}
\title{emln66_temporal_cohen_2003}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   66 \tab Temporal \tab 153 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
