% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln68_temporal_environment_baeta_2011.R
\docType{data}
\name{emln68_temporal_environment_baeta_2011}
\alias{emln68_temporal_environment_baeta_2011}
\title{emln68_temporal_environment_baeta_2011}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   68 \tab Temporal   , Environment \tab 73 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
