% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln69_temporal_fang_2012.R
\docType{data}
\name{emln69_temporal_fang_2012}
\alias{emln69_temporal_fang_2012}
\title{emln69_temporal_fang_2012}
\source{
web

https://figshare.com/articles/dataset/Relative_Stability_of_Core_Groups_in_Pollination_Networks_in_a_Biodiversity_Hotspot_over_Four_Years/127797
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   69 \tab Temporal \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
