% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln6_environment_kohler_2011.R
\docType{data}
\name{emln6_environment_kohler_2011}
\alias{emln6_environment_kohler_2011}
\title{emln6_environment_kohler_2011}
\source{
Web_of_Life

Interaction_Web_DataBase

http://www.web-of-life.es/map.php?type=6

http://www.ecologia.ib.usp.br/iwdb/html/schleuning-et-al-2010.html
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   6 \tab Environment \tab Pollination \tab TRUE \cr
}
}
\keyword{internal}
