% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln70_temporal_mayse_price_1978.R
\docType{data}
\name{emln70_temporal_mayse_price_1978}
\alias{emln70_temporal_mayse_price_1978}
\title{emln70_temporal_mayse_price_1978}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   70 \tab Temporal \tab 158 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
