% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/emln71_temporal_perturbation_kaiser_bunbury_2009.R
\docType{data}
\name{emln71_temporal_perturbation_kaiser_bunbury_2009}
\alias{emln71_temporal_perturbation_kaiser_bunbury_2009}
\title{emln71_temporal_perturbation_kaiser_bunbury_2009}
\source{
Interaction_Web_DataBase

Web_of_Life

http://www.ecologia.ib.usp.br/iwdb/html/kaiser-bunbury_2009.html

http://www.web-of-life.es/map.php
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   71 \tab Temporal    , Perturbation \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
