% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln72_temporal_ponisio_2017.R
\docType{data}
\name{emln72_temporal_ponisio_2017}
\alias{emln72_temporal_ponisio_2017}
\title{emln72_temporal_ponisio_2017}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   72 \tab Temporal \tab 75 \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
