% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln73_temporal_ramos_robles_2016.R
\docType{data}
\name{emln73_temporal_ramos_robles_2016}
\alias{emln73_temporal_ramos_robles_2016}
\title{emln73_temporal_ramos_robles_2016}
\source{
PeerJ

https://peerj.com/articles/2048/#supplemental-information
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   73 \tab Temporal \tab Seed-Dispersal \tab FALSE \cr
}
}
\keyword{internal}
