% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln74_temporal_ruzicka_2012.R
\docType{data}
\name{emln74_temporal_ruzicka_2012}
\alias{emln74_temporal_ruzicka_2012}
\title{emln74_temporal_ruzicka_2012}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   74 \tab Temporal \tab 79 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
