% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln75_temporal_schoenly_reid_1983.R
\docType{data}
\name{emln75_temporal_schoenly_reid_1983}
\alias{emln75_temporal_schoenly_reid_1983}
\title{emln75_temporal_schoenly_reid_1983}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   75 \tab Temporal \tab 149 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
