% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln76_temporal_tavares_cromar_1996.R
\docType{data}
\name{emln76_temporal_tavares_cromar_1996}
\alias{emln76_temporal_tavares_cromar_1996}
\title{emln76_temporal_tavares_cromar_1996}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   76 \tab Temporal \tab 155 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
