% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln77_temporal_vaiela_1974.R
\docType{data}
\name{emln77_temporal_vaiela_1974}
\alias{emln77_temporal_vaiela_1974}
\title{emln77_temporal_vaiela_1974}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   77 \tab Temporal \tab 123 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
