% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln78_temporal_zackenberg_2013.R
\docType{data}
\name{emln78_temporal_zackenberg_2013}
\alias{emln78_temporal_zackenberg_2013}
\title{emln78_temporal_zackenberg_2013}
\source{
plos

https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0081694
}
\description{
Network Description:\tabular{cccc}{
   network_id \tab multilayer_network_type \tab ecological_network_type \tab state_nodes \cr
   78 \tab Temporal \tab Pollination \tab FALSE \cr
}
}
\keyword{internal}
