% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln7_environment_rejmanek_stary_1979.R
\docType{data}
\name{emln7_environment_rejmanek_stary_1979}
\alias{emln7_environment_rejmanek_stary_1979}
\title{emln7_environment_rejmanek_stary_1979}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   7 \tab Environment \tab 112 \tab Host-Parasite \tab FALSE \cr
}
}
\keyword{internal}
