% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln9_environment_seifert_1976.R
\docType{data}
\name{emln9_environment_seifert_1976}
\alias{emln9_environment_seifert_1976}
\title{emln9_environment_seifert_1976}
\source{
mangal

https://mangal.io/doc/api/
}
\description{
Network Description:\tabular{ccccc}{
   network_id \tab multilayer_network_type \tab mangal_code \tab ecological_network_type \tab state_nodes \cr
   9 \tab Environment \tab 117 \tab Food-Web \tab FALSE \cr
}
}
\keyword{internal}
