% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_igraph.R
\name{get_igraph}
\alias{get_igraph}
\title{Convert a multilayer class to a a list of igraph objects}
\usage{
get_igraph(multilayer, bipartite, directed)
}
\arguments{
\item{multilayer}{The multilayer object.}

\item{bipartite}{Is the network bipartite? Must be provided (no default
value).}

\item{directed}{Is the network directed? Must be provided (no default value).}
}
\value{
A list:
\itemize{
  \item \code{layers_igraph} A list of igraph objects, one per layer
  \item \code{nodes} A table of physical nodes, as in the multlayer object.
  \item \code{state_nodes_map} A table with state nodes to map the nodes to the row and columns of a SAM.
}
}
\description{
This function creates a list of igraph objects from the network layers. Interlayer links are not included.
}
\details{
If link, physical node or state node attributes are included in the multilayer network they are passed to the igraph objects.
}
\examples{

# See examples in: https://ecological-complexity-lab.github.io/emln_package/multilayer.html#To_igraph_objects

}
