% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_emln.R
\name{load_emln}
\alias{load_emln}
\title{Load networks include din the package}
\usage{
load_emln(network_id)
}
\arguments{
\item{network_id}{The network ID.}
}
\value{
A multilayer object (see ?multilayer).
}
\description{
Load a specific network by specifying its ID.
}
\details{
Use the \code{search_emln} and \code{view_emln} functions to obtain network IDs.
}
\examples{
\dontrun{
load_emln(network_id = 38)
}
}
\seealso{
\code{multilayer}
}
