% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Internal function - loads an RData file, and returns it}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{file to be downloaded}
}
\description{
Internal function - loads an RData file, and returns it
}
\keyword{internal}
