% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_emln.R
\name{search_emln}
\alias{search_emln}
\title{Search networks}
\usage{
search_emln(
  ecological_network_type = NULL,
  multilayer_network_type = NULL,
  state_nodes = NULL,
  node_number_minimum = NULL,
  layer_number_minimum = NULL,
  weighted = NULL,
  interlayer = NULL,
  directed = NULL
)
}
\arguments{
\item{ecological_network_type}{One of: Seed-Dispersal, Pollination, Food-Web, Host-Parasite, Multiple, Plant-Herbivore, Anemone-Fish, Plant-Ant}

\item{multilayer_network_type}{Environment, Temporal, Multiplex, Perturbation, Spatial}

\item{state_nodes}{TRUE/FALSE}

\item{node_number_minimum}{The minimum number of layers wanted in the network}

\item{layer_number_minimum}{The minimum number of layers wanted in the network}

\item{weighted}{TRUE/FALSE}

\item{interlayer}{TRUE/FALSE}

\item{directed}{TRUE/FALSE}
}
\value{
A tibble containing the network id, and all the search arguments for each of the networks that are compatible with the search.
}
\description{
Search the database for networks depending on search parameters specified below.
}
\examples{
\dontrun{
# See examples in: https://ecological-complexity-lab.github.io/emln_package/data.html#Browsing_the_data

# Generates a tibble of all the networks in the package that are pollination networks
search_emln(ecological_network_type = 'Plant-Ant')

# Generates a tibble of all the networks in the package that have state nodes and are temporal
search_emln(multilayer_network_type = 'Temporal', state_node = TRUE)

}
}
