% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emln.R
\docType{package}
\name{emln-package}
\alias{emln}
\alias{emln-package}
\title{emln: A package to organize, handle, and explore ecological multilayer networks}
\description{
An R package for data and analysis of ecological multilayer networks, including standardization of data structures and functions to convert between them. This work was supported by research grant ISF (Israel Science Foundation) 1281/20 to Shai Pilosof.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ecological-complexity-lab.github.io/emln_package/index.html}
  \item Report bugs at \url{https://github.com/Ecological-Complexity-Lab/emln_package/issues}
}

}
\keyword{internal}
